tools/auto-code-rover/requirements.txt
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-tools==0.2.1
anyio==4.2.0
astroid==3.2.3
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli
certifi==2024.7.4
cffi
cfgv==3.4.0
charset-normalizer
click==8.1.7
coverage==7.5.3
Cython==3.0.8
dill==0.3.8
discord==2.3.2
discord.py==2.3.2
distlib==0.3.8
distro==1.9.0
docker==7.0.0
docstring-parser==0.15
emojis==0.7.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
h2
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.5
hyperframe
icecream==2.1.3
identify==2.5.33
idna==3.7
importlib-metadata==7.0.1
install-jdk==1.1.0
isort==5.13.2
javalang==0.13.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
libclang==18.1.1
linkify-it-py==2.0.2
litellm==1.44.8
loguru==0.7.2
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.4
natsort==8.4.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
ollama==0.3.3
openai==1.50.2
opt-einsum==3.3.0
packaging==23.2
platformdirs==4.1.0
polars==0.20.31
pre-commit==3.6.0
pycparser
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pylint==3.2.3
pyro-api==0.1.2
pyro-ppl==1.9.0
PySocks
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.32.1
regex==2024.5.15
rich==13.7.1
rpds-py==0.16.2
semver==3.0.2
slack_sdk==3.26.2
sniffio==1.3.0
soupsieve==2.5
sympy==1.13.0
tenacity==8.2.3
termcolor==2.4.0
textual==0.52.1
tiktoken==0.7.0
timeout-decorator==0.5.0
tokenizers==0.19.1
tomlkit==0.13.0
torch==2.2.1
tqdm==4.66.4
tree-sitter==0.21.3
tree-sitter-c==0.21.4
tree-sitter-cpp==0.22.2
tree-sitter-java==0.21.0
tree-sitter-languages==1.10.2
triton==2.2.0
types-jsonschema==4.21.0.20240311
typing_extensions==4.12.2
uc-micro-py==1.0.2
unidiff==0.7.5
unittest-xml-reporting==3.2.0
urllib3
virtualenv==20.25.0
yarl==1.9.4
zipp==3.19.2
zstandard==0.22.0
together
javalang
bugswarm-common
